Optimise non-busy-sleep.
This commit amends non-busy sleep to store an absolute tick time for sleeping threads and keep this in an ordered list. To determine which thread(s) we need to wake up on each timer interrupt, we can look loop from the beginning of the list of sleeping threads. If the thread tick value for which it can wake up from is >= to the current tick counter, the thread can be woken up. If not, as the list is ordered, we can stop iterating over the list of sleeping threads.